+++ src/components/Exchange/PositionsList.js
@@ -27,7 +27,7 @@ import { getUsd } from "domain/tokens/utils";
 import { bigNumberify, formatAmount } from "lib/numbers";
 import { AiOutlineEdit } from "react-icons/ai";
 import useAccountType, { AccountType } from "lib/wallets/useAccountType";
-import getLiquidation from "lib/getLiquidation";
+import getLiquidation from "lib/positions/getLiquidation";
 
 const getOrdersForPosition = (account, position, orders, nativeTokenAddress) => {
   if (!orders || orders.length === 0) {
