+++ src/pages/Exchange/Exchange.js
@@ -52,7 +52,7 @@ import { getToken, getTokenBySymbol, getTokens, getWhitelistedTokens } from "con
 import { useChainId } from "lib/chains";
 import ExternalLink from "components/ExternalLink/ExternalLink";
 import UsefulLinks from "components/Exchange/UsefulLinks";
-import { getLeverageNew } from "lib/getLeverageNew";
+import { getLeverageNew } from "lib/positions/getLeverageNew";
 const { AddressZero } = ethers.constants;
 
 const PENDING_POSITION_VALID_DURATION = 600 * 1000;
